Add an abstract type `HypothesisTest`
This is currently defined in HypothesisTests. Defining the common
supertype in this package would allow other packages to define their own
hypothesis test types that are too specific and/or have too many
dependencies to upstream to HypothesisTests. This could include, for
example, the log-rank test for comparing Kaplan-Meier survival curves in
Survival and the likelihood ratio test in StatsModels.